Show all running threads
Sort the table with the slowest command at the top. It is probably more
interesting to follow up the slow commands than the fast ones.

Another possibility is to replace finished rows with new commands, but
then there will just be a mixture of slow and fast ones all over the
place.

Signed-off-by: Fredrik Medley <fredrik.medley@gmail.com>