hash based input checking
when a rule specifies "hash_input = 1" in addition to the
modification time of its inputs the hashed content of the
inputs are taken into account. hence a target is only dirty
if the incoming edges (actually leafes) are dirty content-wise.

the purpose of the change is addressing workflows with a lot of
branch back-and-forth switches. when quickly changing to a
completely different code line, fixing a small issue and switching
back, often a lot of files are touched due to the branch switching.
but as they were not built in between, their artifacts are still
up-to-date but considered dirty by ninja based on the timestamp.
the same holds for other merge or rebase operations.

this is implemented with a hash log that stores filename, mtime
and murmur2 hash of a file into a binary file '.ninja_hashes'.
in addition a combined hash is stored for each successfully biuld target
to reflect the contents of the inputs at the time the output was produced.
this is necessary to also behave correctly for partial builds.

to distinguish between the two different variants a variant discriminator
is persisted with every hash log entry.

the file path is stored (rather than a hash of it) to circumvent
hash collisions which are really problematic as the tool would
not behave deterministic anymore. this problem really exists :-)

the mtime of the hashed file is stored to reduce the number of
hash lookups of a touched but unchanged file. it is only checked once
and subsequent calls compare first of all the mtime of the hashed file.

the syntax is slightly extended, but the existing code-path is unchanged
(besides the two ifs) in case the feature is not enabled. hence no
performance impact on the existing behaviour is expected.