new hash log version
In hash log version 4 paths were persisted as byte sequences terminated by a
NUL byte ahead of the rest of the data.  This meant when a path was read every
byte was read individually and checked for 0.  In hash log version 5 we store
the size of a record followed by the integer data and finally the path string
with a padding.  This improves reading speed because only two read calls are
necessary for each record.