Record mtime of each output, not the largest
Previously, the maximum mtime of all outputs from an edge, or possibly
the most recent input mtime in case of a restat rule, was recorded
in the build log for each output node of an edge. There is no reason
for why, so use the mtime for each individual output node instead.

Signed-off-by: Fredrik Medley <fredrik.medley@gmail.com>